move graphql codegen to nested dir
