+++ src/lib/codegen/graphql/flavors/default/__tests__/examples/directives.simple.ts
@@ -1,4 +1,9 @@
-import { SelectionWrapper } from "@/lib/codegen/flavors/default/wrapper";
+import {
+    SelectionWrapper,
+    SLW_DIRECTIVE,
+    SLW_DIRECTIVE_ARGS,
+    SLW_DIRECTIVE_ARGS_META,
+} from "@/lib/codegen/graphql/flavors/default/wrapper";
 import { makeSLFN, selectScalars } from "../utils";
 import {
     type ArgumentsTypeFromFragment,
@@ -7,11 +12,38 @@ import {
     type SLWsFromSelection,
 } from "../utils/types";
 
-export type ArticleArrayBooksArgs = { language?: string };
-export type ArticleArrayRollDiceArgs = { numDice: number; numSides?: number };
-export type ArticleBooksArgs = { language?: string };
-export type ArticleRollDiceArgs = { numDice: number; numSides?: number };
-export type QuerySearchArgs = { title?: string };
+export type Directive_tagArgs = {
+    /** tag? */
+    tag?: string;
+};
+export type Directive_includeArgs = {
+    /** Included when true. */
+    if: boolean;
+};
+export type Directive_skipArgs = {
+    /** Skipped when true. */
+    if: boolean;
+};
+export type ArticleArrayBooksArgs = {
+    language?: string;
+};
+export type ArticleArrayRollDiceArgs = {
+    numDice: number;
+    numSides?: number;
+};
+export type ArticleBooksArgs = {
+    language?: string;
+};
+export type ArticleRollDiceArgs = {
+    numDice: number;
+    numSides?: number;
+};
+export type QuerySearchArgs = {
+    title?: string;
+};
+export const Directive_tagArgsMeta = { tag: "String" } as const;
+export const Directive_includeArgsMeta = { if: "Boolean!" } as const;
+export const Directive_skipArgsMeta = { if: "Boolean!" } as const;
 export const ArticleArrayBooksArgsMeta = { language: "String" } as const;
 export const ArticleArrayRollDiceArgsMeta = {
     numDice: "Int!",
@@ -23,40 +55,6 @@ export const ArticleRollDiceArgsMeta = {
     numSides: "Int",
 } as const;
 export const QuerySearchArgsMeta = { title: "String" } as const;
-
-export type BookArraySelectionFields = {
-    title?: string;
-    author?: string;
-};
-
-export type ArticleArraySelectionFields = {
-    title?: string;
-    publisher?: string;
-    books: (args?: ArticleArrayBooksArgs) => typeof BookArraySelection;
-    rollDice: (args: ArticleArrayRollDiceArgs) => Array<number>;
-};
-
-export type SearchResultArraySelectionFields =
-    | BookSelectionFields
-    | ArticleSelectionFields;
-
-export type BookSelectionFields = {
-    title?: string;
-    author?: string;
-};
-
-export type ArticleSelectionFields = {
-    title?: string;
-    publisher?: string;
-    books: (args?: ArticleBooksArgs) => typeof BookArraySelection;
-    rollDice: (args: ArticleRollDiceArgs) => Array<number>;
-};
-
-export type QuerySelectionFields = {
-    books: typeof BookArraySelection;
-    articles: typeof ArticleArraySelection;
-    search: (args?: QuerySearchArgs) => typeof SearchResultArraySelection;
-};
 
 export function makeBookArraySelectionInput(this: any) {
     return {
@@ -78,6 +76,7 @@ export function makeBookArraySelectionInput(this: any) {
             }) as (
                 ...args: ArgumentsTypeFromFragment<F>
             ) => ReturnTypeFromFragment<F>,
+
         $scalars: () =>
             selectScalars(
                 makeBookArraySelectionInput.bind(this)(),
@@ -131,6 +130,7 @@ export function makeArticleArraySelectionInput(this: any) {
             }) as (
                 ...args: ArgumentsTypeFromFragment<F>
             ) => ReturnTypeFromFragment<F>,
+
         $scalars: () =>
             selectScalars(
                 makeArticleArraySelectionInput.bind(this)(),
@@ -189,6 +189,7 @@ export function makeBookSelectionInput(this: any) {
             }) as (
                 ...args: ArgumentsTypeFromFragment<F>
             ) => ReturnTypeFromFragment<F>,
+
         $scalars: () =>
             selectScalars(
                 makeBookSelectionInput.bind(this)(),
@@ -240,6 +241,7 @@ export function makeArticleSelectionInput(this: any) {
             }) as (
                 ...args: ArgumentsTypeFromFragment<F>
             ) => ReturnTypeFromFragment<F>,
+
         $scalars: () =>
             selectScalars(
                 makeArticleSelectionInput.bind(this)(),
@@ -255,59 +257,68 @@ export const ArticleSelection = makeSLFN(
     0,
 );
 
-export function makeQuerySelectionInput(this: any) {
+type ReturnTypeFromQuerySelection = {
+    books: ReturnType<
+        SLFN<
+            {},
+            ReturnType<typeof makeBookArraySelectionInput>,
+            "BookArraySelection",
+            "Book",
+            1,
+            {
+                $lazy: () => Promise<"T">;
+            },
+            "$lazy"
+        >
+    >;
+    articles: ReturnType<
+        SLFN<
+            {},
+            ReturnType<typeof makeArticleArraySelectionInput>,
+            "ArticleArraySelection",
+            "Article",
+            1,
+            {
+                $lazy: () => Promise<"T">;
+            },
+            "$lazy"
+        >
+    >;
+    search: (args: QuerySearchArgs) => ReturnType<
+        SLFN<
+            {},
+            ReturnType<typeof makeSearchResultArraySelectionInput>,
+            "SearchResultArraySelection",
+            "SearchResult",
+            1,
+            {
+                $lazy: (args: QuerySearchArgs) => Promise<"T">;
+            },
+            "$lazy"
+        >
+    >;
+} & {
+    $fragment: <F extends (this: any, ...args: any[]) => any>(
+        f: F,
+    ) => (...args: ArgumentsTypeFromFragment<F>) => ReturnTypeFromFragment<F>;
+};
+
+export function makeQuerySelectionInput(
+    this: any,
+): ReturnTypeFromQuerySelection {
     return {
-        books: BookArraySelection.bind({
-            collector: this,
-            fieldName: "books",
-        }) as ReturnType<
-            SLFN<
-                {},
-                ReturnType<typeof makeBookArraySelectionInput>,
-                "BookArraySelection",
-                "Book",
-                1,
-                {
-                    $lazy: () => Promise<"T">;
-                },
-                "$lazy"
-            >
-        >,
+        books: BookArraySelection.bind({ collector: this, fieldName: "books" }),
         articles: ArticleArraySelection.bind({
             collector: this,
             fieldName: "articles",
-        }) as ReturnType<
-            SLFN<
-                {},
-                ReturnType<typeof makeArticleArraySelectionInput>,
-                "ArticleArraySelection",
-                "Article",
-                1,
-                {
-                    $lazy: () => Promise<"T">;
-                },
-                "$lazy"
-            >
-        >,
+        }),
         search: (args: QuerySearchArgs) =>
             SearchResultArraySelection.bind({
                 collector: this,
                 fieldName: "search",
                 args,
                 argsMeta: QuerySearchArgsMeta,
-            }) as ReturnType<
-                SLFN<
-                    {},
-                    ReturnType<typeof makeSearchResultArraySelectionInput>,
-                    "SearchResultArraySelection",
-                    "SearchResult",
-                    1,
-                    {
-                        $lazy: (args: QuerySearchArgs) => Promise<"T">;
-                    },
-                    "$lazy"
-                >
-            >,
+            }),
 
         $fragment: <F extends (this: any, ...args: any[]) => any>(f: F) =>
             f.bind({
@@ -317,10 +328,6 @@ export function makeQuerySelectionInput(this: any) {
             }) as (
                 ...args: ArgumentsTypeFromFragment<F>
             ) => ReturnTypeFromFragment<F>,
-        $scalars: () =>
-            selectScalars(
-                makeQuerySelectionInput.bind(this)(),
-            ) as SLWsFromSelection<ReturnType<typeof makeQuerySelectionInput>>,
     } as const;
 }
 export const QuerySelection = makeSLFN(
@@ -330,22 +337,45 @@ export const QuerySelection = makeSLFN(
     0,
 );
 
-export type _RootOperationSelectionFields<T extends object> = {
-    query: ReturnType<
-        SLFN<
-            T,
-            ReturnType<typeof makeQuerySelectionInput>,
-            "QuerySelection",
-            "Query",
-            0
-        >
-    >;
-};
+export const _directive_tag =
+    (args: Directive_tagArgs) =>
+    <F>(f: F) => {
+        (f as any)[SLW_DIRECTIVE] = "tag";
+        (f as any)[SLW_DIRECTIVE_ARGS] = args;
+        (f as any)[SLW_DIRECTIVE_ARGS_META] = Directive_tagArgsMeta;
+        return f;
+    };
+
+export const _directive_include =
+    (args: Directive_includeArgs) =>
+    <F>(f: F) => {
+        (f as any)[SLW_DIRECTIVE] = "include";
+        (f as any)[SLW_DIRECTIVE_ARGS] = args;
+        (f as any)[SLW_DIRECTIVE_ARGS_META] = Directive_includeArgsMeta;
+        return f;
+    };
+
+export const _directive_skip =
+    (args: Directive_skipArgs) =>
+    <F>(f: F) => {
+        (f as any)[SLW_DIRECTIVE] = "skip";
+        (f as any)[SLW_DIRECTIVE_ARGS] = args;
+        (f as any)[SLW_DIRECTIVE_ARGS_META] = Directive_skipArgsMeta;
+        return f;
+    };
+
+export const $directives = {
+    tag: _directive_tag,
+    include: _directive_include,
+    skip: _directive_skip,
+} as const;
 export function _makeRootOperationInput(this: any) {
     return {
         query: QuerySelection.bind({
             collector: this,
             isRootType: "Query",
         }),
+
+        $directives,
     } as const;
 }
